vest_lib

Module buf_traits

Definitions for buffer traits that can be used as input and output for parsers and serializers, along with some implementations for commonly used buffers.

Re-exports

pub use crate::utils::*;

Traits

VestInput

Trait definitions Trait for types that can be used as input for Vest parsers, roughly corresponding to byte buffers. VestInput does not expose the contents of the buffer, so opaque buffer types for side-channel security can implement VestInput.

VestOutput

Trait for types that can be used as output for Vest serializers. VestOutput does not expose the contents of the buffer, so opaque buffer types for side-channel security can implement VestOutput.

VestPublicInput

Trait for types that can be used as input for Vest parsers, roughly corresponding to byte buffers. VestPublicInput can be set using transparent bytes, so it cannot provide type abstraction for side-channel security.

VestPublicOutput

Trait for types that can be used as output for Vest serializers. VestPublicOutput can be set using transparent bytes, so it cannot provide type abstraction for side-channel security.