vest_lib::buf_traits

Trait VestPublicInput

pub trait VestPublicInput: VestInput {
    // Required method
    exec fn as_byte_slice(&self) -> res : &[u8] ;
}

Trait for types that can be used as input for Vest parsers, roughly corresponding to byte buffers. VestPublicInput can be set using transparent bytes, so it cannot provide type abstraction for side-channel security.

Required Methods

exec fn as_byte_slice(&self) -> res : &[u8]

ensures
res@ == self@,

Returns a byte slice with the contents of the buffer