vest_lib::combinators::recursive::spec

Trait SpecRecBody

pub trait SpecRecBody {
    type Param;
    type T;
    type Body: SpecCombinator<T = Self::T>;

    // Required method
    spec fn spec_body(
        &self,
        param: Self::Param,
        rec: ParamRecSpecs<Self::Param, Self::T>,
    ) -> Self::Body;
}

Defines one level of a recursive format for use with super::FixWith.

Required Associated Types

type Param

type T

type Body: SpecCombinator<T = Self::T>

Required Methods

spec fn spec_body( &self, param: Self::Param, rec: ParamRecSpecs<Self::Param, Self::T>, ) -> Self::Body

Define one recursive unfolding for param, where rec provides callbacks for all recursive positions in the body.