vest_lib::combinators::star

Struct Star

pub struct Star<A>(pub A);

Kleene star combinator: greedy zero-or-more repetition, consuming/producing Seq<A::PVal>.

Parsing semantics: always succeeds (may return an empty sequence). Stops when A fails or consumes zero bytes.

Consistency

A sequence s is consistent with Star<A> iff every element of s is consistent with A.

Note

This combinator is mostly used internally to specify Repeat<A, B>, which is able to disambiguate A and B and hence more compositional.

Tuple Fields
0: A
Implementations

impl<A: SpecParser> Star<A>

pub open spec fn parse_rec(&self, ibuf: Seq<u8>) -> (int, Seq<A::PVal>)

{
    match self.0.spec_parse(ibuf) {
        Some((n, v)) if 0 < n <= ibuf.len() => {
            let (n_rest, vs) = self.parse_rec(ibuf.skip(n));
            (n + n_rest, seq![v] + vs)
        }
        _ => (0, Seq::empty()),
    }
}

Recursive helper function for parsing. Since Star always succeeds, this function is total.

impl<A: SpecByteLen> Star<A>

pub proof fn lemma_byte_len_cons(&self, v: A::T, vs: Seq<A::T>)

ensures
self.byte_len(seq![v] + vs) == self.0.byte_len(v) + self.byte_len(vs),

impl<A: SpecSerializer + Copy> Star<A>

pub proof fn lemma_spec_serialize_cons(&self, first: A::SVal, rest: Seq<A::SVal>)

ensures
self.spec_serialize(seq![first] + rest)
    == self.0.spec_serialize(first) + self.spec_serialize(rest),

Serializing a nonempty sequence is the serialization of its first value followed by the serialization of the remaining values.

pub proof fn lemma_spec_serialize_suffix_step(&self, vs: Seq<A::SVal>, i: int)

requires
0 <= i < vs.len(),
ensures
self.spec_serialize(vs.skip(i))
    == self.0.spec_serialize(vs[i]) + self.spec_serialize(vs.skip(i + 1)),

Decomposes the serialization suffix beginning at i.

impl<A: SpecSerializerDps> Star<A>

pub open spec fn rfold_serialize_dps(&self, vs: Seq<A::SValue>, obuf: Seq<u8>) -> Seq<u8>

{ vs.fold_right_alt(|elem, buf| self.0.spec_serialize_dps(elem, buf), obuf) }