vest_lib::core::exec::serializer

Trait Serializer

pub trait Serializer<Output, T>
where
    Output: OutputBuf,
    Self: SpecByteLen<T = T::V> + SpecSerializer<SVal = T::V> + Consistency<Val = T::V>,
    T: DeepView + ?Sized,
{
    // Required method
    exec fn serialize_into(&self, v: &T, obuf: &mut Output);

    // Provided method
    open spec fn exec_inv(&self) -> bool { ... }
}

An executable serializer targeting Output.

Required Methods

exec fn serialize_into(&self, v: &T, obuf: &mut Output)

requires
self.exec_inv(),

self.consistent(v.deep_view()),

old(obuf).fits(self.byte_len(v.deep_view())),
ensures
final(obuf)@ == old(obuf)@ + self.spec_serialize(v.deep_view()),

forall |n| old(obuf).fits(self.byte_len(v.deep_view()) + n) <==> final(obuf).fits(n),

old(obuf).same_destination(final(obuf)),

Serializes the value v into the output buffer obuf by (logically) appending the serialized bytes to the end of obuf. This view has two main benefits:

It matches the specification of the serializer closely, which simplifies the proofs;

It is general enough to support both fixed-size (e.g., &mut [u8]) and growable (e.g., Vec<u8>) output buffers (see OutputBuf).

Preconditions

The serializer’s execution invariant holds (mainly used for super::fns::FnSerializer, usually trivial for most combinators).

The value v is compliant with the format specification.

The output buffer has enough space to hold the serialized value.

Postconditions

The output buffer’s contents are extended by the serialized value.

The output buffer’s remaining capacity is reduced by the serialized value’s length.

The output buffer’s destination remains the same.

Provided Methods

open spec fn exec_inv(&self) -> bool

{ true }