vest_lib::core::spec

Trait Consistency

pub trait Consistency {
    type Val;

    // Required method
    spec fn consistent(&self, v: Self::Val) -> bool;
}

Value well-formedness according to a format’s specification.

Examples

crate::combinators::Refined requires the refinement predicate;

crate::combinators::Varied/crate::combinators::Repeat requires matching length/count;

crate::combinators::U8 imposes trivial consistency condition (all u8 values are consistent with the format);

crate::combinators::Void is uninhabited and thus has no consistent values.

Required Associated Types

type Val

The type of values whose consistency is being checked.

Required Methods

spec fn consistent(&self, v: Self::Val) -> bool

Returns true if v is well-formed w.r.t. this combinator.