vest_lib::core::spec

Trait SpecByteLen

pub trait SpecByteLen {
    type T;

    // Required method
    spec fn byte_len(&self, v: Self::T) -> nat;
}

Denotes the byte length of a value w.r.t. a combinator’s format spec.

Required Associated Types

type T

The type of values whose byte length is being computed.

Required Methods

spec fn byte_len(&self, v: Self::T) -> nat

Returns the number of bytes v occupies when serialized.