vest_lib

Module properties

Definitions for parser and serializer combinators and their properties.

Re-exports

pub use crate::buf_traits::*;

pub use crate::errors::*;

Traits

Combinator

Implementation for parser and serializer combinators. A combinator’s view must be a SecureSpecCombinator.

SecureSpecCombinator

Theorems and lemmas that must be proven for a combinator to be considered correct and secure.

SpecCombinator

Specification for parser and serializer Combinators. All Vest combinators must implement this trait.

Type Aliases

PResult

The parse result of a combinator.

SResult

The serialize result of a combinator.