vest_lib::properties

Trait SpecCombinator

pub trait SpecCombinator {
    type Type;

    // Required methods
    spec fn spec_parse(&self, s: Seq<u8>) -> Option<(int, Self::Type)>;
    spec fn spec_serialize(&self, v: Self::Type) -> Seq<u8>;

    // Provided methods
    open spec fn wf(&self, v: Self::Type) -> bool { ... }
    fn requires(&self) -> bool { ... }
}

Specification for parser and serializer Combinators. All Vest combinators must implement this trait.

Required Associated Types

type Type

The view of [Combinator::Result].

Required Methods

spec fn spec_parse(&self, s: Seq<u8>) -> Option<(int, Self::Type)>

The specification of Combinator::parse.

spec fn spec_serialize(&self, v: Self::Type) -> Seq<u8>

The specification of Combinator::serialize.

Provided Methods

open spec fn wf(&self, v: Self::Type) -> bool

{ true }

Well-formedness of the format [Self::type] (e.g., refinements on the type).

open spec fn requires(&self) -> bool

{ true }

Pre-conditions for parsing and serialization.

Examples

Sequencing combinators require that the first combinator is prefix-secure.

Repetition combinators require that the inner combinator is prefix-secure.

crate::regular::repetition::Repeat combinator requires that the inner combinator is productive.

crate::regular::variant::Choice combinator requires that Snd is crate::regular::disjoint::DisjointFrom Fst.