vest_lib::regular::disjoint

Trait DisjointFrom

pub trait DisjointFrom<Other>
where
    Self: SpecCombinator,
    Other: SpecCombinator,
{
    // Required methods
    spec fn disjoint_from(&self, other: &Other) -> bool;
    proof fn parse_disjoint_on(&self, other: &Other, buf: Seq<u8>);
}

A helper trait for Choice combinator.

Required Methods

spec fn disjoint_from(&self, other: &Other) -> bool

pre-condition that must be held for Self and [Other] to be disjoint

proof fn parse_disjoint_on(&self, other: &Other, buf: Seq<u8>)

requires
self.disjoint_from(other),
ensures
self.spec_parse(buf) is Some ==> other.spec_parse(buf) is None,

The combinator Self is disjoint from the combinator [Other] if the bytes buf can be parsed by Self but not by [Other]