vest_lib::regular

Module modifier

Combinator modifiers

Structs

AndThen

Combinator that monadically chains two combinators.

Cond

Combinator that checks if cond is true and then delegates to the inner combinator.

Mapped

Combinator that maps the result of an inner combinator with an isomorphism that implements Iso.

Refined

Combinator that refines the result of an inner combinator with a predicate that implements Pred.

TryMap

Combinator that maps the result of an inner combinator with a fallible conversion that implements [TryFromInto].

Traits

Iso

All isomorphisms to be used in Mapped combinator must implement this trait.

IsoFn

The bijective functions of Iso Self::apply and Self::rev_apply must be inverses of each other. See [SpecIsoFn::spec_iso] and [SpecIsoFn::spec_iso_rev] for more details.

PartialIso

Fallible version of Iso.

PartialIsoFn

The fallible functions of PartialIso

Pred

All predicates to be used in Refined combinator must implement this trait.

SpecIso

Spec version of Iso.

SpecIsoFn

The bijective functions of SpecIso

SpecIsoProof

The proof of SpecIsoFn

SpecPartialIso

Spec version of PartialIso.

SpecPartialIsoFn

The fallible functions of SpecPartialIso

SpecPartialIsoProof

The proof of SpecPartialIsoFn

SpecPred

The spec version of Pred.