vest_lib::regular::sequence

Struct Pair

pub struct Pair<Fst, Snd, Cont> {
    pub fst: Fst,
    pub _snd: PhantomData<Snd>,
    pub snd: Cont,
}

Combinator that sequentially applies two combinators, where the second combinator depends on the first one.

Fields
fst: Fst

combinators that contain dependencies

_snd: PhantomData<Snd>

phantom data representing the second combinator (it really should be private, but this is a workaround for Verus’s conservative treatment of opaque types, which doesn’t allow field access of opaque types in open spec functions)

snd: Cont

closure that captures dependencies and maps them to the dependent combinators

Implementations

impl<Fst, Snd, Cont> Pair<Fst, Snd, Cont>
where Fst: View, Snd: View, Cont: View<V = GhostFn<<Fst::V as SpecCombinator>::Type, Snd::V>>, Fst::V: SecureSpecCombinator, Snd::V: SpecCombinator,

pub exec fn new(fst: Fst, snd: Cont) -> o : Self

ensures
o.fst == fst,

o.snd == snd,

o@ == Pair::spec_new(fst@, snd@),

Creates a new Pair combinator.

impl<Fst, Snd> Pair<Fst, Snd, GhostFn<Fst::Type, Snd>>
where Fst: SecureSpecCombinator, Snd: SpecCombinator,

pub open spec fn spec_new(fst: Fst, snd: GhostFn<Fst::Type, Snd>) -> Self

{
    Pair {
        fst,
        _snd: core::marker::PhantomData,
        snd,
    }
}

Creates a new Pair combinator.