vest_lib::regular::sequence

Type Alias SpecPair

pub type SpecPair<Fst, Snd> = Pair<Fst, Snd, GhostFn<<Fst as SpecCombinator>::Type, Snd>>;

Alias for a spec dependent pair combinator.

Aliased Type

pub struct SpecPair<Fst, Snd> {
    pub fst: Fst,
    pub _snd: PhantomData<Snd>,
    pub snd: FnSpec<(<Fst as SpecCombinator>::Type,), Snd>,
}

Fields
fst: Fst

combinators that contain dependencies

_snd: PhantomData<Snd>

phantom data representing the second combinator (it really should be private, but this is a workaround for Verus’s conservative treatment of opaque types, which doesn’t allow field access of opaque types in open spec functions)

snd: FnSpec<(<Fst as SpecCombinator>::Type,), Snd>

closure that captures dependencies and maps them to the dependent combinators