vest_lib::regular::uints

Trait FromToBytes

pub trait FromToBytes
where
    Self: ViewReflex + Sized + Copy,
{

    // Required methods
    spec fn spec_from_le_bytes(s: Seq<u8>) -> Self;
    spec fn spec_to_le_bytes(&self) -> Seq<u8>;
    spec fn spec_from_be_bytes(s: Seq<u8>) -> Self;
    spec fn spec_to_be_bytes(&self) -> Seq<u8>;
    proof fn lemma_spec_to_from_le_bytes(&self);
    proof fn lemma_spec_from_to_le_bytes(s: Seq<u8>);
    proof fn lemma_spec_from_le_bytes_no_lookahead(s1: Seq<u8>, s2: Seq<u8>);
    proof fn lemma_spec_to_from_be_bytes(&self);
    proof fn lemma_spec_from_to_be_bytes(s: Seq<u8>);
    proof fn lemma_spec_from_be_bytes_no_lookahead(s1: Seq<u8>, s2: Seq<u8>);
    exec fn ex_from_le_bytes(s: &[u8]) -> x : Self;
    exec fn ex_to_le_bytes<I, O>(&self, s: &mut O, pos: usize)
       where I: VestPublicInput,
             O: VestPublicOutput<I>;
    exec fn ex_from_be_bytes(s: &[u8]) -> x : Self;
    exec fn ex_to_be_bytes<I, O>(&self, s: &mut O, pos: usize)
       where I: VestPublicInput,
             O: VestPublicOutput<I>;
    exec fn eq(&self, other: &Self) -> res : bool;

}

A trait for converting an integer type to and from a sequence of bytes.

Required Methods

spec fn spec_from_le_bytes(s: Seq<u8>) -> Self

Spec version of Self::ex_from_le_bytes

spec fn spec_to_le_bytes(&self) -> Seq<u8>

Spec version of Self::ex_to_le_bytes

spec fn spec_from_be_bytes(s: Seq<u8>) -> Self

Spec version of Self::ex_from_be_bytes

spec fn spec_to_be_bytes(&self) -> Seq<u8>

Spec version of Self::ex_to_be_bytes

proof fn lemma_spec_to_from_le_bytes(&self)

ensures
self.spec_to_le_bytes().len() == size_of::<Self>(),

Self::spec_from_le_bytes(self.spec_to_le_bytes()) == *self,

Lemma that asserts that converting an integer in little-endian byte order to bytes and back results in the original integer

proof fn lemma_spec_from_to_le_bytes(s: Seq<u8>)

ensures
s.len() == size_of::<Self>()
    ==> Self::spec_to_le_bytes(&Self::spec_from_le_bytes(s)) == s,

Lemma that asserts that converting a sequence of bytes to an integer in little-endian byte order and back results in the original sequence

proof fn lemma_spec_from_le_bytes_no_lookahead(s1: Seq<u8>, s2: Seq<u8>)

ensures
s1.len() >= size_of::<Self>()
    ==> Self::spec_from_le_bytes(s1) == Self::spec_from_le_bytes(s1.add(s2)),

Helper lemma for proving SecureSpecCombinator::lemma_prefix_secure

proof fn lemma_spec_to_from_be_bytes(&self)

ensures
self.spec_to_be_bytes().len() == size_of::<Self>(),

Self::spec_from_be_bytes(self.spec_to_be_bytes()) == *self,

Lemma that asserts that converting an integer in big-endian byte order to bytes and back results in the original integer

proof fn lemma_spec_from_to_be_bytes(s: Seq<u8>)

ensures
s.len() == size_of::<Self>()
    ==> Self::spec_to_be_bytes(&Self::spec_from_be_bytes(s)) == s,

Lemma that asserts that converting a sequence of bytes to an integer in big-endian byte order and back results in the original sequence

proof fn lemma_spec_from_be_bytes_no_lookahead(s1: Seq<u8>, s2: Seq<u8>)

ensures
s1.len() >= size_of::<Self>()
    ==> Self::spec_from_be_bytes(s1) == Self::spec_from_be_bytes(s1.add(s2)),

Helper lemma for proving SecureSpecCombinator::lemma_prefix_secure

exec fn ex_from_le_bytes(s: &[u8]) -> x : Self

requires
s@.len() == size_of::<Self>(),
ensures
x == Self::spec_from_le_bytes(s@),

Converts a sequence of bytes to an integer in little-endian byte order.

exec fn ex_to_le_bytes<I, O>(&self, s: &mut O, pos: usize)
where I: VestPublicInput, O: VestPublicOutput<I>,

requires
old(s)@.len() - pos >= size_of::<Self>(),
ensures
old(s)@.len() == s@.len(),

self.spec_to_le_bytes().len() == size_of::<Self>(),

s@ == seq_splice(old(s)@, pos, self.spec_to_le_bytes()),

Converts an integer to a sequence of bytes in little-endian byte order.

exec fn ex_from_be_bytes(s: &[u8]) -> x : Self

requires
s@.len() == size_of::<Self>(),
ensures
x == Self::spec_from_be_bytes(s@),

Converts a sequence of bytes to an integer in big-endian byte order.

exec fn ex_to_be_bytes<I, O>(&self, s: &mut O, pos: usize)
where I: VestPublicInput, O: VestPublicOutput<I>,

requires
old(s)@.len() - pos >= size_of::<Self>(),
ensures
old(s)@.len() == s@.len(),

self.spec_to_be_bytes().len() == size_of::<Self>(),

s@ == seq_splice(old(s)@, pos, self.spec_to_be_bytes()),

Converts an integer to a sequence of bytes in big-endian byte order.

exec fn eq(&self, other: &Self) -> res : bool

ensures
res == (self@ == other@),

Compares two integers for equality.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.