vest_lib

Module utils

Utility functions and types.

Traits

Compare

A helper trait for two different types that can be compared.

From

Vest equivalent of std::convert::From.

Into

Vest equivalent of std::convert::Into.

SpecFrom

Spec version of From.

SpecInto

Spec version of Into.

SpecTryFrom

Spec version of TryFrom.

SpecTryInto

Spec version of TryInto.

TryFrom

Vest equivalent of std::convert::TryFrom.

TryInto

Vest equivalent of std::convert::TryInto.

ViewReflex

Helper trait for types that have a reflexive view.

Functions

compare_slice

init_vec_u8

seq_splice

set_range

vec_u8_extend_from_slice